# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "SecureAuth MFA OTP & CAPTCHA API",
        "description": "Complete API documentation for the SecureAuth application.\n\n## Running modes\n\n### Monolith mode (`npm run start`)\n- Single server on `:3000` implements registration (OTP → CAPTCHA → TOTP setup) and login (password → TOTP).\n\n### Multi-server mode (`npm run start:multi`)\n- **OTP/UI** server on `:3000`\n- **CAPTCHA** server on `:5000`\n- **MFA** server on `:8000`\n\nIn multi-server mode, login is **password → login-OTP → CAPTCHA → MFA**.\n\n## Flows\n\n### Registration (monolith & multi-server OTP server)\n1. `POST /api/register` — Register & send OTP (email or SMS)\n2. `POST /api/verify-email` — Verify the 6-digit OTP\n3. `GET /api/captcha` — Get SVG CAPTCHA challenge (registration)\n4. `POST /api/verify-captcha` — Verify CAPTCHA & get TOTP QR code\n5. Scan QR code in Google Authenticator\n\n### Login (multi-server)\n1. `POST /api/login` — Password check, then send login OTP\n2. `POST /api/verify-login-otp` — Verify login OTP\n3. `GET /api/captcha` — Get SVG CAPTCHA challenge (login CAPTCHA server)\n4. `POST /api/verify-captcha` — Verify CAPTCHA (login CAPTCHA server)\n5. `POST /api/verify-mfa` — Verify TOTP (MFA server)\n\n### Session\n- `GET /api/me` — Get current logged-in user\n- `POST /api/logout` — Destroy session",
        "version": "1.1.0",
        "contact": {
            "name": "SecureAuth",
...
        }
    },
    "servers": [
        {
...
        {
...
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Register user & send OTP",
                "description": "Register a new user and send a 6-digit OTP to their email address or phone number via SMS.\n\n**DEV MODE:** If `EMAIL_USER`/`EMAIL_PASS` are not set in `.env`, the OTP is logged to the server console instead of being emailed. Similarly for Twilio SMS.",
                "operationId": "registerUser",
                "servers": [
                    {
                        "url": "http://localhost:{otpPort}",
                        "description": "OTP/UI server only",
                        "variables": {
                            "otpPort": {
                                "default": "3000",
                                "description": "Your OTP/UI port (e.g. 3500)."
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            },
                            "examples": {
                                "email": {
                                    "summary": "Register with Email",
                                    "value": {
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "phone": "",
                                        "password": "SecurePass123!",
                                        "method": "email"
                                    }
                                },
                                "phone": {
                                    "summary": "Register with Phone (SMS)",
                                    "value": {
                                        "name": "Jane Doe",
                                        "email": "",
                                        "phone": "+1234567890",
                                        "password": "SecurePass456!",
                                        "method": "phone"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponse"
                                },
                                "example": {
                                    "success": true,
                                    "identifier": "john@example.com",
                                    "message": "Verification email sent."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email or phone already registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Email already registered."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to send email/SMS",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Failed to send email after multiple attempts. Check configuration."
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login with credentials",
                "description": "Authenticate a registered user with their email/phone and password.\n\n- In **monolith mode**, this returns `requireMfa: true`.\n- In **multi-server mode**, this returns `requireLoginOtp: true` and sends a login OTP, then you must call `POST /api/verify-login-otp` before going to the CAPTCHA service.",
                "operationId": "loginUser",
                "servers": [
                    {
                        "url": "http://localhost:{otpPort}",
                        "description": "OTP/UI server only",
                        "variables": {
                            "otpPort": {
                                "default": "3000",
                                "description": "Your OTP/UI port (e.g. 3500)."
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            },
                            "example": {
                                "identifier": "john@example.com",
                                "password": "SecurePass123!"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Credentials valid — next step required (depends on mode)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                },
                                "example": {
                                    "success": true,
                                    "requireLoginOtp": true,
                                    "message": "Password valid. Please enter the OTP sent to your email/phone."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Invalid credentials."
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/verify-mfa": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify MFA (TOTP code)",
                "description": "Verify the 6-digit TOTP code from Google Authenticator (or any TOTP-compatible app).\n\n- Monolith mode: requires `identifier` and verifies it against the user's stored secret.\n- Multi-server mode: verifies the code for the *pending login user* stored in session, and requires previous steps (login OTP + CAPTCHA) to be completed.\n\n**Note:** TOTP codes rotate every 30 seconds.",
                "operationId": "verifyMfa",
                "servers": [
                    {
                        "url": "http://localhost:3000",
                        "description": "Monolith server"
                    },
                    {
                        "url": "http://localhost:8000",
                        "description": "MFA service (multi-server)"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyMfaRequest"
                            },
                            "example": {
                                "identifier": "john@example.com",
                                "mfaCode": "847291"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA verified — user is logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerifyMfaResponse"
                                },
                                "example": {
                                    "success": true,
                                    "user": {
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "phone": ""
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "User not found."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid Authenticator code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Invalid Authenticator code."
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/me": {
            "get": {
                "tags": [
                    "Session"
                ],
                "summary": "Get current logged-in user",
                "description": "Returns the profile of the currently logged-in user. Requires an active session (session cookie).",
                "operationId": "getCurrentUser",
                "servers": [
                    {
                        "url": "http://localhost:{otpPort}",
                        "description": "OTP/UI server only",
                        "variables": {
                            "otpPort": {
                                "default": "3000",
                                "description": "Your OTP/UI port (e.g. 3500)."
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User is logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                },
                                "example": {
                                    "user": {
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "phone": ""
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "error": "Not logged in."
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RegisterRequest": {
                "type": "object",
                "required": [
                    "name",
                    "password",
                    "method"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "User's full name",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Email address (required when method = 'email')",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Phone number with country code (required when method = 'phone')",
                        "example": "+1234567890"
                    },
                    "password": {
                        "type": "string",
                        "description": "User password",
                        "example": "SecurePass123!"
                    },
                    "method": {
                        "type": "string",
                        "enum": [
                            "email",
                            "phone"
                        ],
                        "description": "Verification method — 'email' or 'phone'",
                        "example": "email"
                    }
                }
            },
            "RegisterResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "identifier": {
                        "type": "string",
                        "description": "The email or phone used as the user identifier",
                        "example": "john@example.com"
                    },
                    "message": {
                        "type": "string",
                        "example": "Verification email sent."
                    }
                }
            },
...
            "LoginRequest": {
                "type": "object",
                "required": [
                    "identifier",
                    "password"
                ],
                "properties": {
                    "identifier": {
                        "type": "string",
                        "description": "Registered email or phone",
                        "example": "john@example.com"
                    },
                    "password": {
                        "type": "string",
                        "description": "User password",
                        "example": "SecurePass123!"
                    }
                }
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "requireMfa": {
                        "type": "boolean",
                        "description": "Monolith mode: indicates MFA verification is required",
                        "example": true
                    },
                    "requireLoginOtp": {
                        "type": "boolean",
                        "description": "Multi-server mode: indicates login-OTP verification is required next",
                        "example": true
                    },
                    "devOtp": {
                        "type": "string",
                        "description": "Present only when `SHOW_DEV_OTPS=true` (multi-server).",
                        "example": "123456"
                    },
                    "message": {
                        "type": "string",
                        "example": "Password valid. Please enter the OTP sent to your email/phone."
                    }
                }
            },
            "VerifyMfaRequest": {
                "type": "object",
                "required": [
                    "mfaCode"
                ],
                "properties": {
                    "identifier": {
                        "type": "string",
                        "description": "Monolith mode only: registered email or phone. Multi-server mode uses the pending login identifier from session.",
                        "example": "john@example.com"
                    },
                    "mfaCode": {
                        "type": "string",
                        "description": "6-digit TOTP code from Authenticator app",
                        "pattern": "^[0-9]{6}$",
                        "example": "847291"
                    }
                }
            },
...
            "VerifyMfaResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                }
            },
            "UserResponse": {
                "type": "object",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": ""
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message",
                        "example": "Something went wrong."
                    }
                }
            }
        }
    }
}